Change stuff as discussed with @sheeep.
